Added VFX Graph
Didn't really do much here, just added VFX Graph along with some doors for the Main Floor of the house. I have a desk prefab that I went and added a box collider to and textured and untextured face. I also went and changed the rotation of the desk in my two scenes.